refactorize metadata
