+++ src/dev/lab/lab-constants.ts
@@ -1,0 +1,11 @@
+import { show } from './util'
+
+(async () => {
+	const context = { }
+
+	const list = [
+		// 'a = true',
+		'a = nvl(null,false)'
+	]
+	show(list, context)
+})()
