+++ src/dev/lab/lab.ts
@@ -1,5 +1,5 @@
 import { expressions as exp } from '../../lib'
 (async () => {
-	const parameters = exp.parameters('a + 1')
-	console.log(JSON.stringify(parameters))
+	exp.addFunction((n1: number, n2: number): number => n1 % n2, 'xxx(n1:number,n2:number):number')
+	console.log(exp.eval('xxx(7,2)'))
 })()
