+++ src/lib/model/index.ts
@@ -3,5 +3,4 @@ export * from './operands'
 export * from './cache'
 export * from './observer'
 export * from './managers'
-export * from './parser'
 export * from './type'
