+++ src/lib/operand/index.ts
@@ -1,5 +1,4 @@
 export * from './operands'
 export * from './operandBuilder'
-export * from './library'
 export * from './serializer'
 export * from './typeManager'
