+++ src/lib/parser/index.ts
@@ -1,3 +1,3 @@
 export * from './expressionConfig'
 export * from './node'
-export * from './parserManager'
+export * from './parser'
